requirements.txt
streamlit==1.39.0
plotly==5.24.0
pandas==2.2.2
Pillow==10.3.0
